Implementing the core layout.
The OpenNMS servers configuration has been implemented. This is the core
of the application, as the rest of the information presented to the
user, depends on the chosen server.

The SLM View, as well as the events table have been implemented as well.

The events view has server side capabilities, and infinite scroll based
on the pagination features of the OpenNMS ReST API. The search results
require tuning due to current ReST limitations.

The rest of the view are just placeholders for now.