Refactoring Service Layer
All the services will be independent of the selected OpenNMS server.

The appropriate handling of the chosen server will be managed inside the
new HttpService class. This class will subscribe to an EventEmitter that
should be responsible for notify every time the default server is
changed. This emiter will be implemented inside the OnmsServersService
class.

The components have been simplified as none of them depend on the
OnmsServersService class.

The appropriate registration is handled inside the main component.

A new error handling has been implemented, but it has to be properly
updated when the remote server is unresponsive.