New Features: App Badges and Summary Charts
When the app goes to background (pause), a badge with the current number
of alarms is displayed on the app icon if the user provides permissions
for this.

The alarm and outage summary now have charts similar to Compass.

The package.json was updated to reflect the required native plugins.